+++ tests/resources/configs/iso19115_0_standard.py
@@ -667,6 +667,7 @@ complete_record_v4 = {
         "spatial_resolution": 1000000,
         "lineage": {"statement": "Example lineage statement"},
         "edition": "2",
+        "series": {"title": {"value": "Test Records"}, "edition": "2024 Series"},
         "other_citation_details": "Author, A., Author, B., & Author, C. (2022). The title (Version 1.0) [Data set]. Publisher. https://doi.org/the-doi",
         "identifiers": [
             {